services/scraper/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
httpx>=0.27
websockets>=12.0
python-multipart>=0.0.9
playwright>=1.45
google-generativeai>=0.8.3
anthropic>=0.97
pillow>=10.0
python-dotenv>=1.0
